refactor(settings): split TOGGLE_DEFS out of the toggle store
settingKeys.ts only needs the toggle definitions, and importing the
zustand store for them broke any test that partially mocks
toggleSettingsStore (collection error: TOGGLE_DEFS missing from the
mock). toggleSettingsStore re-exports unchanged for its 52 existing
importers.
